Make DhcpOption own its `data` memory.
Rather than borrowing a `[u8]`, let the `data` field of `DhcpOption` be
a `Vec<u8>`. This allows `DhcpOption` values to be generated on the fly
without imposing lifetime requirements on the code generating the
values; it also removes the lifetime from `Packet`, allowing packets to
be created and moved more freely.